@@ -23,10 +23,14 @@
     "@angular/platform-browser": "^5.2.0",
     "@angular/platform-browser-dynamic": "^5.2.0",
     "@angular/router": "^5.2.0",
+    "@ngrx/effects": "^6.1.0",
+    "@ngrx/store": "^6.1.0",
+    "@ngrx/store-devtools": "^6.1.0",
     "core-js": "^2.4.1",
     "hammerjs": "^2.0.8",
     "rxjs": "^6.2.2",
     "rxjs-compat": "^6.2.2",
+    "save": "^2.3.2",
     "zone.js": "^0.8.19"
   },
   "devDependencies": {
@@ -50,6 +54,6 @@
     "sass-loader": "^7.0.1",
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
-    "typescript": "~2.5.3"
+    "typescript": "~2.9.2"
   }
 }
