@@ -1,20 +1,19 @@
 import {
   Component,
   OnInit,
-  Input,
-  OnChanges,
   OnDestroy,
 } from '@angular/core';
 
-import { Subscription, ReplaySubject, pipe } from 'rxjs';
-import { switchMap } from 'rxjs/operators';
+import { Store } from '@ngrx/store';
+import { AppState } from '../../app.state';
+import * as fromStore from '../../core/store';
+
+import { Subscription, ReplaySubject, Observable } from 'rxjs';
 
 import { MatDialog } from '@angular/material';
 import { MatDialogRef } from '@angular/material';
 
-import { Course } from '../../shared/classes/course.class';
-import { CourseItem } from '../../shared/interfaces/course.interface';
-import { CourseData } from '../../shared/interfaces/course-data.interface';
+import { CourseItem, CourseData } from '../../shared/interfaces';
 import { CourseService } from '../../core/services/course.service';
 import { OnDeleteDialogComponent } from './on-delete-dialog/on-delete-dialog.component';
 import { CourseSearchPipe } from '../../core/pipes/course-search.pipe';
@@ -27,10 +26,11 @@ import { CourseSearchPipe } from '../../core/pipes/course-search.pipe';
     CourseSearchPipe,
   ]
 })
-export class CourseListComponent implements OnInit, OnChanges, OnDestroy {
+export class CourseListComponent implements OnInit, OnDestroy {
   private pageCount: number;
-  public coursesList: Array<CourseItem> = [];
-  public filteredList: Array<CourseItem> = [];
+  private store: Store<AppState>;
+
+  public coursesList$: Observable<CourseItem[]>;
   public dialog: MatDialog;
   public dialogRef: MatDialogRef<OnDeleteDialogComponent>;
   public courseSearchPipe: CourseSearchPipe;
@@ -38,52 +38,39 @@ export class CourseListComponent implements OnInit, OnChanges, OnDestroy {
   public courseSubscriber: Subscription;
   public deleteCourseSubscription: Subscription;
   public dynamicCourseUpload: ReplaySubject<number> = new ReplaySubject();
-
-  @Input('searchValue')
-  public searchValue: string;
+  public searchValue: string = '';
+  public storedCourses: Observable<CourseItem[]>;
 
   constructor(
     courseService: CourseService,
     dialog: MatDialog,
     courseSearchPipe: CourseSearchPipe,
+    store: Store<AppState>
   ) {
     this.courseService = courseService;
     this.courseSearchPipe = courseSearchPipe;
     this.dialog = dialog;
-  }
-
-  private createCourseItem(element: CourseData): CourseItem {
-    return new Course(element.id, element.name, element.date, element.length, element.description, element.isTopRated);
+    this.store = store;
   }
 
   public ngOnInit(): void {
-    this.getCourses();
-    this.pageCount = 2;
-    this.courseService.courseSubject
-      .subscribe(data => this.coursesList = data.map(element => {
-        return this.createCourseItem(element);
-      }));
-    this.dynamicCourseUpload.pipe(
-      switchMap(pageCount => this.courseService.getCourses(pageCount)))
-      .subscribe((data) => {
-        this.coursesList = this.coursesList.concat(data.map(element => this.createCourseItem(element)));
+    this.coursesList$ = this.store.select(fromStore.getAllCourses);
+    this.courseService.courseSubject.subscribe(query => {
+      this.searchValue = query;
+      this.store.dispatch(new fromStore.SearchCourse(query));
     });
-  }
-
-  public ngOnChanges(): void {
-    this.filteredList = this.courseSearchPipe.transform(this.coursesList, this.searchValue);
+    this.store.dispatch(new fromStore.LoadCourses(true));
+    this.pageCount = 2;
   }
 
   public ngOnDestroy(): void {
     if (this.deleteCourseSubscription) {
       this.deleteCourseSubscription.unsubscribe();
     }
-    this.dynamicCourseUpload.unsubscribe();
-    this.courseSubscriber.unsubscribe();
   }
 
   public onCourseUploadClick(): void {
-    this.dynamicCourseUpload.next(Number(this.pageCount));
+    this.store.dispatch(new fromStore.LoadCourses(Number(this.pageCount)));
     this.pageCount += 1;
   }
 
@@ -91,16 +78,8 @@ export class CourseListComponent implements OnInit, OnChanges, OnDestroy {
     this.dialogRef = this.dialog.open(OnDeleteDialogComponent, {height: '10rem', width: '10rem'});
     this.deleteCourseSubscription = this.dialogRef.afterClosed().subscribe(result => {
       if (result) {
-        this.courseService.removeItem(event)
-        .then(() => this.getCourses());
+        this.store.dispatch(new fromStore.DeleteCourse(event));
       }
     });
-  }
-
-  public getCourses(): void {
-    this.courseSubscriber = this.courseService.getCourses(this.pageCount)
-      .subscribe(data => this.coursesList = data.map(element => {
-        return this.createCourseItem(element);
-      }));
   }
 }
