@@ -5,7 +5,7 @@ import { Router } from '@angular/router';
 import { CourseService } from '../../core/services/course.service';
 import { validateDate } from '../../core/validators/date-validator';
 import { ROUTES_CONFIG } from '../../../../app-config/routes/routes.config';
-import { CourseData } from '../../shared/interfaces/course-data.interface';
+import { CourseData } from '../../shared/interfaces';
 
 @Component({
   selector: 'add-course-form',
