@@ -1,13 +1,15 @@
 import { Pipe, PipeTransform } from '@angular/core';
-import { CourseItem } from '../../shared/interfaces/course.interface';
+import { CourseItem } from '../../shared/interfaces';
 
 @Pipe({name: 'sortByDate'})
 
 export class SortByDatePipe implements PipeTransform {
   public transform(list: Array<CourseItem>): Array<CourseItem> {
-    list.sort((first, second) => {
-      return Date.parse(second.date.toString()) - Date.parse(first.date.toString());
-    });
-    return list;
+    if (list) {
+      list.sort((first, second) => {
+        return Date.parse(second.date.toString()) - Date.parse(first.date.toString());
+      });
+      return list;
+    }
   }
 }
