@@ -3,18 +3,20 @@ import { HttpClient, HttpParams } from '@angular/common/http';
 
 import { ReplaySubject } from 'rxjs';
 import { Observable } from 'rxjs';
+import { map, catchError } from 'rxjs/operators';
+import { of } from 'rxjs/observable/of';
 
 import { COURSELIST } from '../../shared/mocks/mock-courses';
-import { CourseItem } from '../../shared/interfaces/course.interface';
-import { CourseData } from '../../shared/interfaces/course-data.interface';
+import { CourseItem, CourseData } from '../../shared/interfaces';
 import { BASE_URL } from '../../shared/constants/path-config';
+import { Course } from '../../shared/classes/course.class';
 
 @Injectable()
 export class CourseService {
   private http: HttpClient;
   private baseUrl: string = BASE_URL;
   public state: string;
-  public courseSubject: ReplaySubject<Array<CourseData>> = new ReplaySubject();
+  public courseSubject: ReplaySubject<string> = new ReplaySubject();
 
   constructor (
     http: HttpClient,
@@ -25,19 +27,42 @@ export class CourseService {
 
   public getCourses(pageCount: number): Observable<CourseData[]> {
     const urlParams: HttpParams = new HttpParams().set('_limit', '5').set('_page', pageCount + '');
-    return this.http.get<any>(`${this.baseUrl}/courses/courses/courses`, {params: urlParams});
+    return this.http.get<any>(`${this.baseUrl}/courses/courses/courses`, {params: urlParams}).pipe(
+      map(courses => {
+        return courses.map(element => new Course(
+          element.id,
+          element.name,
+          element.date,
+          element.length,
+          element.description,
+          element.isTopRated));
+      }),
+      catchError(error => of(error))
+    );
   }
 
-  public removeItem(id: number): Promise<any> {
-    return this.http.delete(`${this.baseUrl}/courses/${id}`).toPromise();
+  public removeItem(id: number): Observable<any> {
+    return this.http.delete(`${this.baseUrl}/courses/${id}`);
   }
 
-  public serverCourseSearch(query: string): void {
-    const urlParams: HttpParams = new HttpParams().set('name_like', query);
-    this.http.get(`${this.baseUrl}/courses/courses/courses`, {params: urlParams})
-        .subscribe((data: CourseData[]) => {
-          this.courseSubject.next(data);
-        });
+  public onSearchAction(query: string): void {
+    this.courseSubject.next(query);
+  }
+
+  public serverCourseSearch(query: string): Observable<any> {
+    const urlParams: HttpParams = new HttpParams().set('name_like', query);    
+    return this.http.get(`${this.baseUrl}/courses/courses/courses`, {params: urlParams}).pipe(
+      map((courses: CourseData[]) => {        
+        return courses.map(element => new Course(
+          element.id,
+          element.name,
+          element.date,
+          element.length,
+          element.description,
+          element.isTopRated));
+      }),
+      catchError(error => of(error))
+    );
   }
 
   public getSpecificCourse(id: string): Observable<{}> {
