@@ -1,6 +1,9 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { StoreModule } from '@ngrx/store';
+import { EffectsModule } from '@ngrx/effects';
 
+import { reducers, effects } from '../core/store';
 import { HomePageComponent } from './home.component';
 import { HeaderModule } from '../shared/header';
 import { FooterModule } from '../shared/footer';
@@ -13,7 +16,9 @@ import { ToolboxModule } from './toolbox/toolbox.module';
     HeaderModule,
     FooterModule,
     CourseListModule,
-    ToolboxModule
+    ToolboxModule,
+    StoreModule.forFeature('courses', reducers),
+    EffectsModule.forFeature(effects)
   ],
   declarations: [
     HomePageComponent
